worker/requirements.txt
# Message Queue
celery==5.6.2
pika==1.3.2
amqp==5.3.1
kombu==5.6.2
billiard==4.2.4
vine==5.1.0
redis==6.4.0

# HTTP Client
httpx==0.28.1

# Config
pydantic-settings==2.13.1
pydantic==2.12.5
python-dotenv==1.2.2

# Video Download
yt-dlp==2026.3.17

# Speech-to-Text
openai-whisper==20250625

# Text-to-Speech (뉴스 내레이션)
edge-tts>=6.1.0

# Video Editing
moviepy==2.2.1
ffmpeg-python==0.2.0

# Object Storage
minio==7.2.20
